Add Redis client configuration to main.go
Added Redis client configuration in the main function of the main.go file in the service directory. The configuration options include details like address, password, and DB, which are retrieved from environment variables if available. Furthermore, the Redis client has been passed as an argument to the cache test handler.